Recreated Visual C 6 projects and workspace.
Needed to go back to VC6 to get to a working baseline before trying to
debug the VC2010 versions. Here I created 3 VC6 projects and a
workspace to hold them.

NOTE: The VC6 and VC2010 projects coexist in the same directories but
they are not 100% compatible at the level of the intermediate compiler
and linker files. If you compile in VC6 then switch to VC2010 you MUST
rebuild everything. Similarly, if you have compiled in VC2010 but
return to work in VC6, you MUST rebuild the projects.