Move CodaRTnetSDK outside of repository.
CodaRTnetSDK is now a subdirectory of “Code Projects”. The version
number is removed from the directory, but the version currently stored
at that location should be found in “CodaRTnetSDK/CodaRTnetSDK
VersionInfo.rtf”. This way the C++ files can reference a generic
directory. The desired version of the SDK should be copied into that
location.